+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/annotation/builders/SVCNVAnnotationBuilder.java
@@ -1,0 +1,5 @@
+package de.charite.compbio.jannovar.annotation.builders;
+
+public class SVCNVAnnotationBuilder {
+
+}